requirements.txt
# AI Economy Protocol (AEP) - Dependencies

# Stage 1: ASI Alliance Foundation
uagents>=0.12.0
uagents-core>=0.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Stage 2: Sanctum Gateway Integration
solana>=0.30.2
solders>=0.18.1
anchorpy>=0.18.0
base58>=2.1.1
httpx>=0.24.0
websockets>=11.0
# x402 Integration
flask[async]>=3.0.0
flask-cors>=4.0.0
asgiref>=3.7.0
nest-asyncio>=1.6.0

# AI Integration (Gemini)
google-generativeai>=0.3.0
aiohttp>=3.9.0
